pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
include = ["**/*.py","**/*.pkl"]

[project]
name = "mrotools"
version = "2.0.0.0"
license = "MIT"
description = "SNR Tools"
authors = [
  { name="Dr. Eros Montin, PhD", email="eros.montin@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["boto3","matplotlib","scipy","Pillow","cloudmrhub-tools @ git+https://github.com/cloudmrhub/cloudmr-tools.git","raider_eros_montin @ git+https://github.com/erosmontin/raider.git","pynico_eros_montin @ git+https://github.com/erosmontin/pynico.git","pyable_eros_montin @ git+https://github.com/erosmontin/pyable.git"]

[project.urls]
"Homepage" = "https://github.com/cloudmrhub/mroptimum-tools"
"Bug Tracker" = "https://github.com/cloudmrhub/mroptimum-tools/issues"
